backend/requirements.txt
flask
flask-cors
cryptography
dnspython
requests
python-dotenv
psutil
colorama
sqlalchemy
pyjwt
reportlab
